re-use orderbook package to calculate rounding slippage

+++ services/horizon/internal/config.go
@@ -95,6 +95,6 @@ type Config struct {
 	// balances like ELB or ALB. In such case http.Request.RemoteAddr will be
 	// replaced with the last IP in X-Forwarded-For header.
 	BehindAWSLoadBalancer bool
-	// RoundingSlippageFilter excludes trades from /trade_aggregations with rounding slippage >x%
-	RoundingSlippageFilter float64
+	// RoundingSlippageFilter excludes trades from /trade_aggregations with rounding slippage >x bps
+	RoundingSlippageFilter uint
 }
